LPS-98441 Move com.liferay.portal.security.auth integration tests to module
