+++ modules/apps/portal-security/portal-security-auth-test/src/testIntegration/java/com/liferay/portal/security/auth/test/DefaultScreenNameGeneratorTest.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.security.auth;
+package com.liferay.portal.security.auth.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.security.auth.DefaultScreenNameGenerator;
 import com.liferay.portal.kernel.security.auth.ScreenNameGenerator;
@@ -26,11 +27,13 @@ import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
  * @author Riccardo Ferrari
  * @author Daniel Reuther
  */
+@RunWith(Arquillian.class)
 public class DefaultScreenNameGeneratorTest {
 
 	@ClassRule
